% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_utils.R
\name{create_data.table_n_rep_prior}
\alias{create_data.table_n_rep_prior}
\title{Create data.table objects}
\usage{
create_data.table_n_rep_prior(n_vec, n_rep, prior_vec, covariates)
}
\arguments{
\item{n_vec}{Array of numbers of observations}

\item{n_rep}{Number of repetitions to done of the simulation}

\item{prior_vec}{Array of prior names to compare}

\item{covariates}{Array with covariate names}
}
\description{
Create data.table objects
}
