% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{get_corr_data}
\alias{get_corr_data}
\title{Generate correlated data}
\usage{
get_corr_data(p, x = rnorm(100), corr_mat, seed)
}
\arguments{
\item{p}{Number of correlated covariates}

\item{x}{Generate covariates correlated to `x`.}

\item{corr_mat}{Correlation matix, has to be of dim `p`. If missing
it uses a random correlation matrix.}

\item{seed}{Integer to use as random seed. If missing, no seed is set.}
}
\description{
Generate correlated data
}
\examples{
get_corr_data(10, seed=7)
}
