% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{get_posdef_matrix}
\alias{get_posdef_matrix}
\title{Generating a random positive-definite matrix}
\usage{
get_posdef_matrix(n, ev = runif(n, 0, 10))
}
\arguments{
\item{n}{Dimension of the matrix}

\item{ev}{eigenvalue vector}
}
\description{
Generating a random positive-definite matrix with user-specified
positive eigenvalues. If eigenvalues are not specified, they are
generated from a uniform distribution.
}
\details{
Function taken from https://stat.ethz.ch/pipermail/r-help/2008-February/153708
}
