% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_utils.R
\name{get_prior}
\alias{get_prior}
\title{Convert prior code to mobmf priorspec object}
\usage{
get_prior(priorname, n)
}
\arguments{
\item{priorname}{One of the following strings: mom, normid, zell, gzell, gmom}

\item{n}{Number of observations in data where prior is to be used}
}
\description{
Automatically creates priorspec objects from a string. For mom `tau=0.34`;
for normid and groupmom `tau=1` and for zellner and groupzellner `tau=n`
}
