% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_utils.R
\name{get_prior_pair}
\alias{get_prior_pair}
\title{Convert prior pair code to priorspec objects}
\usage{
get_prior_pair(pair_name, n)
}
\arguments{
\item{pair_name}{`_` separated prior codes.}

\item{n}{Number of observations}
}
\description{
Convert a `_` separated pair of prior codes (as defined in [get_prior()])
}
\seealso{
[get_prior()]
}
