% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_utils.R
\name{update_data.table}
\alias{update_data.table}
\title{Store fit results in data.tables}
\usage{
update_data.table(model_dt, marg_dt, row, fit, model_idxs, covariates)
}
\arguments{
\item{model_dt}{Data table to store model probability and model frequency}

\item{marg_dt}{Data table to store marginal inclusion probabilities}

\item{row}{Row of the data.table to modify}

\item{fit}{Result of [mombf::modelSelection()]}

\item{model_idxs}{String containing the indexs of the model whose details have to be stored}

\item{covariates}{Array with covariate names}
}
\description{
Store fit results in data.tables
}
